+++ R/init.R
@@ -1,4 +1,4 @@
-#' @importFrom graphics image.default image par plot title box text axis plot.new plot.window rasterImage layout lcm
+#' @importFrom graphics image.default image par plot title box text axis plot.new plot.window rasterImage layout lcm contour
 #' @importFrom grDevices dev.capabilities dev.size grey rgb
 #' @importFrom utils head tail setTxtProgressBar txtProgressBar
 #' @importFrom stats na.omit runif aggregate setNames predict